+++ src/components/form/index.ts
@@ -1,0 +1,2 @@
+export { Form, FormActions } from './Form'
+export type { FormActionsProps, FormProps } from './Form'